requirements.txt
wheel
toml
mmc-export
mdutils
MdUtils
PyYAML
requests-oauthlib